+++ packages/nimble-components/src/icon-base/types.ts
@@ -2,6 +2,7 @@
  * Predefined icon status states
  * @public
  */
+export type IconStatusAttribute = 'fail' | 'warning' | 'pass' | 'regular';
 export enum IconStatus {
     Fail = 'fail',
     Warning = 'warning',