+++ packages/nimble-components/src/text-field/types.ts
@@ -1,0 +1,9 @@
+import { TextFieldType } from '@microsoft/fast-foundation';
+
+export type TextFieldTypeAttribute =
+    | 'email'
+    | 'password'
+    | 'tel'
+    | 'text'
+    | 'url';
+export { TextFieldType };