+++ packages/nimble-components/src/theme-provider/design-tokens.ts
@@ -30,7 +30,7 @@ import {
     GroupLabel1Weight,
     LegacyContentFamily
 } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
-import { NimbleTheme } from './themes';
+import { NimbleTheme } from './types';
 
 function rgbString(hexValue: string): string {
     const { red, green, blue } = hexRgb(hexValue);